Changed the navigation tab order, highlighted the user section
changed the navigation back to the original order. I also changed the
tab for the logged in user section so we could at least see it for our
discussion. The logged in user section should really be a new tab at the
end of the navigation, not buried under the public topic.