Label styles for FORMS
I applied the label style to the forms that I intended to be there all
along with the "required" style (orange) for the profile page. I did add
code to all of the forms -- simply <label></label>. And added <label
class="required"></label> for the required ones. If this causes any
trouble just roll back this commit.